% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{balance}
\alias{balance}
\title{Balance a data.frame}
\usage{
balance(x, n_ind = min(table(x[[1]])), replace = TRUE)
}
\arguments{
\item{x}{data.frame with first column as grouping column}

\item{n_ind}{integer number of individuals in each group. Smallest sample size by default.}

\item{replace}{logical whether so sample with replacement (or not). `TRUE` by default.}
}
\value{
balanced data.frame
}
\description{
Balance with `n` individuals in each group, taken from the first column
}
\details{
Note that [sample]() is done with `replace=TRUE`
}
\examples{
# untouched dataset is not balanced
table(pig$sp)
# balance helps
table(balance(pig)$sp)
}
