% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{lda1}
\alias{lda1}
\title{Linear discricriminant analysis}
\usage{
lda1(x)
}
\arguments{
\item{x}{data.frame}
}
\value{
confusion matrix (of class `table`)
}
\description{
A thin wrapper around `MASS::lda` that returns a (named) confusion matrix
}
\examples{
pig \%>\% lda1()
}
