% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{acc}
\alias{acc}
\alias{acc_classes}
\alias{acc_all}
\title{Confusion matrix metrics}
\usage{
acc(x)

acc_classes(x)

acc_all(x)
}
\arguments{
\item{x}{table typically a confusion matrix}
}
\value{
a named vector
}
\description{
Only accuracy (global and by class) so far but more may come.
}
\examples{
x <- iris \%>\% lda1
x \%>\% acc()
x \%>\% acc_class()
x \%>\% acc_all()
}
