% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{partition}
\alias{partition}
\title{Obtain different flavours of a dataset}
\usage{
partition(x, n_ind = min(table(x[[1]])))
}
\arguments{
\item{x}{data.frame}

\item{n_ind}{passed to [balance()] (the smallest sample size by default)}
}
\value{
a named list described above
}
\description{
This returns a list with four variations of a data.set:
}
\details{
* original: the dataset passed, untouched
* random:   original but shuffled on first (ie grouping) column
* balanced: original but balanced on first column
* balanced_random: balanced but randomized on first column
}
\examples{
pig \%>\% partition()
}
