% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pig.R
\docType{data}
\name{pig}
\alias{pig}
\title{Data from JAS paper}
\format{
a data.frame containing grouping factor as first column and the first 50 principal components
}
\source{
<https://doi.org/10.1016/j.jas.2012.08.005>
}
\usage{
pig
}
\description{
Data from JAS paper
}
\keyword{datasets}
