% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{randomize}
\alias{randomize}
\title{Randomize a data.frame}
\usage{
randomize(x)
}
\arguments{
\item{x}{data.frame with first column as grouping column}
}
\value{
shuffled data.frame
}
\description{
Randomize a data.frame by shuffling the first column
}
\examples{
# untouched dataset
head(pig$sp, 10)
# balance helps
head(randomize(pig)$sp, 10)
}
