% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{retain}
\alias{retain}
\title{Retain some variables from a data.frame}
\usage{
retain(x, n = ncol(x) - 1)
}
\arguments{
\item{x}{data.frame with first column as grouping column}

\item{n}{integer number of columns to retain. All by default.}
}
\value{
data.frame
}
\description{
Retain only the `n` columns, the first let apart.
}
\examples{
# untouched dataset with all its columns
ncol(pig)
# after retaining only 12 (plus the 1st as the grouping)
ncol(retain(pig, 12))
}
